Build warning/error fixes

* remotes/origin/dev/fix-build-warnings-errors:
  Added TODO to change the use of a convenience
  Adjusted test epsilon to pass on Mac machines to 1e-13
  Bumped required Boost version to 1.48
  add static cast type conversions to fix windows warnings
  Reposition function implementation to eliminate warning

+++ maptk/core/geo_reference_points_io.cxx
@@ -107,7 +107,7 @@ void load_reference_file(path_t const& reference_file,
   }
 
   // Initialize lgcs center
-  mean /= reference_lms.size();
+  mean /= static_cast<double>(reference_lms.size());
   lgcs.set_utm_origin(mean);
   cerr << "[load_reference_file] mean position (lgcs origin): " << mean << endl;
